Move http/error folder so the package name matches the import
