+++ integrations/github/github.go
@@ -3,7 +3,7 @@ package github
 import (
 	"fmt"
 	gh "github.com/google/go-github/github"
-	"github.com/weaveworks/flux/http/error"
+	"github.com/weaveworks/flux/http/httperror"
 	"golang.org/x/oauth2"
 	"net/http"
 )
